Add new conversions, refactor
