+++ chia-primitives/src/condition.rs
@@ -1,6 +1,4 @@
 use chia_bls::{SecretKey, Signature};
-use clvm_utils::Allocate;
-use clvmr::{allocator::NodePtr, Allocator};
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub enum Condition {
@@ -9,27 +7,6 @@ pub enum Condition {
         amount: u64,
         memos: Vec<[u8; 32]>,
     },
-}
-
-impl Allocate for Condition {
-    fn from_clvm(_a: &Allocator, _node: NodePtr) -> clvm_utils::Result<Self> {
-        todo!()
-    }
-    fn to_clvm(&self, a: &mut Allocator) -> clvm_utils::Result<NodePtr> {
-        match self {
-            Self::CreateCoin {
-                puzzle_hash,
-                amount,
-                memos,
-            } => {
-                if memos.is_empty() {
-                    Allocate::to_clvm(&(51, (*puzzle_hash, (*amount, ()))), a)
-                } else {
-                    Allocate::to_clvm(&(51, (*puzzle_hash, (*amount, (memos.clone(), ())))), a)
-                }
-            }
-        }
-    }
 }
 
 pub fn sign_agg_sig_me(
