+++ chia-primitives/src/primitives/standard_puzzle.rs
@@ -5,6 +5,7 @@ use clvmr::{allocator::NodePtr, Allocator};
 use crate::{condition::Condition, puzzles::P2_DELEGATED_OR_HIDDEN_HASH};
 
 #[derive(Debug, Clone, ToClvm, FromClvm)]
+#[clvm(curried_args)]
 pub struct StandardPuzzle {
     pub synthetic_key: PublicKey,
 }
