+++ clvm-utils/src/lib.rs
@@ -1,18 +1,18 @@
 pub use clvm_macros::*;
 
+mod convert;
 mod curry;
 mod curry_tree_hash;
 mod error;
 mod new_list;
-mod traits;
 mod tree_hash;
 mod uncurry;
 
+pub use convert::*;
 pub use curry::*;
 pub use curry_tree_hash::*;
 pub use error::*;
 pub use new_list::*;
-pub use traits::*;
 pub use tree_hash::*;
 pub use uncurry::*;
